refactor: UI設定一元化・フック分割・品質向上
