+++ app/layout.tsx
@@ -1,6 +1,8 @@
-import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
+
 import { ClerkProvider } from '@clerk/nextjs'
+
+import type { Metadata } from "next";
 import "./globals.css";
 
 const geistSans = Geist({
