+++ components/layout/Header.tsx
@@ -1,13 +1,17 @@
 'use client'
 
-import { useAuth, UserButton } from '@clerk/nextjs'
 import Link from "next/link"
+
+import { useAuth, UserButton } from '@clerk/nextjs'
+
 import { Navigation } from '@/components/domains/navigation'
+import { useMenuState, useKeyboardNavigation } from '@/lib/domains/navigation/hooks'
+import { BUTTON_CLASSES } from '@/lib/domains/navigation/styles'
+import { useOutsideClick } from '@/lib/shared/hooks'
+
 import { MenuButton } from './MenuButton'
 import { MobileMenu } from './MobileMenu'
-import { useMenuState, useKeyboardNavigation } from '@/lib/domains/navigation/hooks'
-import { useOutsideClick } from '@/lib/shared/hooks'
-import { BUTTON_CLASSES } from '@/lib/domains/navigation/styles'
+
 
 export default function Header() {
   const { isSignedIn } = useAuth()
