+++ eslint.config.mjs
@@ -11,6 +11,59 @@ const compat = new FlatCompat({
 
 const eslintConfig = [
   ...compat.extends("next/core-web-vitals", "next/typescript"),
+  ...compat.extends("plugin:import/recommended"),
+  {
+    rules: {
+      // Import順序の統一ルール
+      "import/order": [
+        "error",
+        {
+          groups: [
+            "builtin",     // Node.js組み込みモジュール
+            "external",    // 外部ライブラリ
+            "internal",    // 内部モジュール（@/パス）
+            "parent",      // 親ディレクトリ
+            "sibling",     // 同一ディレクトリ
+            "index",       // index.js
+            "type",        // 型のみのインポート
+          ],
+          pathGroups: [
+            {
+              pattern: "react",
+              group: "external",
+              position: "before",
+            },
+            {
+              pattern: "next/**",
+              group: "external",
+              position: "before",
+            },
+            {
+              pattern: "@/**",
+              group: "internal",
+              position: "before",
+            },
+          ],
+          pathGroupsExcludedImportTypes: ["react", "next"],
+          "newlines-between": "always",
+          alphabetize: {
+            order: "asc",
+            caseInsensitive: true,
+          },
+        },
+      ],
+      
+      // 使用されていないimportを警告
+      "no-unused-vars": "off",
+      "@typescript-eslint/no-unused-vars": [
+        "warn",
+        { 
+          argsIgnorePattern: "^_",
+          varsIgnorePattern: "^_",
+        },
+      ],
+    },
+  },
 ];
 
 export default eslintConfig;
