+++ lib/domains/navigation/hooks/index.ts
@@ -1,2 +1,5 @@
 export * from './useMenuState'
 export * from './useKeyboardNavigation'
+export * from './useFocusManagement'
+export * from './useKeyboardEvents'
+export * from './useEscapeHandler'
