+++ lib/domains/navigation/hooks/useMenuState.ts
@@ -1,4 +1,5 @@
 import { useState, useCallback } from 'react'
+
 import { MenuState } from '../types'
 import { announceMenuToggle } from '../utils'
