+++ lib/domains/navigation/styles.ts
@@ -1,5 +1,5 @@
-import { cn } from '@/lib/shared/utils/ui'
 import { APP_CONFIG } from '@/lib/shared/config'
+import { cn } from '@/lib/shared/utils/ui'
 
 // ナビゲーション固有スタイルクラス（設定値は config.ts に移動）
